# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Test",
        "description": "Esta es una API de prueba para mostrar a las empresas y poder ser contratado. Espero que le guste :)",
        "contact": {
            "email": "lanlion000128@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar Usuario",
                "description": "Elimina un usuario",
                "operationId": "892733a82d62df86b0731adbc2db048a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "example": "2"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Elimina el usuario satisfactoriamente"
                    },
                    "401": {
                        "description": "Unahutorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unahutorized"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not Authorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Not Authorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Not Found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}